refactor(assistant): inline load_skill into rendering-tools
