+++ apps/studio/lib/ai/tools/index.ts
@@ -7,7 +7,6 @@ import { getIncidentTools } from './incident-tools'
 import { getMcpTools } from './mcp-tools'
 import { getSchemaTools } from './schema-tools'
 import { getRenderingTools } from './rendering-tools'
-import { getSkillTools } from './skill-tools'
 
 export const getTools = async ({
   projectRef,
@@ -24,8 +23,8 @@ export const getTools = async ({
   accessToken?: string
   baseUrl?: string
 }) => {
-  // Always include rendering tools and skill tools
-  let tools: ToolSet = { ...getRenderingTools(), ...getSkillTools() }
+  // Always include rendering tools
+  let tools: ToolSet = getRenderingTools()
 
   // If self-hosted, only add fallback tools
   if (!IS_PLATFORM) {
